Add a module that loads GA credentials on a rotation
This commit adds a function that provides secret keys and report emails
on rotation. This is necessary because we have to maintain several sets
of DAP credentials to stay under Google's rate limits.

This commit lets the analytics reporter call a function to get a set of
creds which are provided on rotation. This means an even dispersment of
creds amongst API calls.

The creds are loaded by looking for environment variables with the
`ANALYTICS_KEY` and `ANALYTICS_REPORT_EMAIL` prefix and then grouping
those based on suffix. So, for example, `ANALYTICS_KEY_1` and
`ANALYTICS_REPORT_EMAIL_1` would be grouped together. Once the creds are
grouped, they are stored in an array. When this new function is called,
it returns the element that succeeds the last element it returned. At
the end of the array, it goes back to the beginning.